switch to using `_fast` integers where possible

+++ include/stdlib.h
@@ -5,8 +5,8 @@
 #define __need_wint_t
 #define __need_NULL
 #include <stddef.h>
-char* itoa(long value, char* result, int base);
-char* utoa(unsigned long value, volatile char* result, int base);
+char *itoa(uint_fast64_t value, char* result, uint_fast8_t base);
+char *utoa(uint_fast64_t value, volatile char* result, uint_fast8_t base);
 void *malloc(size_t size);
 void *realloc(void *ptr, size_t size);
 void free(void *ptr);
